Trying to get subreports working

+++ WpfApplication3/ViewModel/PrintInvoiceLineViewModel.cs
@@ -10,7 +10,7 @@ namespace WpfApplication3.ViewModel
     public class PrintInvoiceLineViewModel : ViewModelBase
     {
         public string Roba { get; set; }
-        public int Amount { get; set; }
-        public int Price { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Price { get; set; }
     }
 }
